requirements.txt
aiohappyeyeballs==2.4.4
aiohttp>=3.9.1,<3.10.0
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
box==0.1.5
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Columnar==1.3.1
dataclasses-json==0.6.7
distro==1.9.0
eval_type_backport==0.2.2
executing==0.8.2
filelock==3.17.0
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.38.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-community==0.3.16
langchain-core==0.3.32
langchain-together==0.3.0
langsmith==0.3.1
loguru==0.7.3
markdown-it-py==3.0.0
marshmallow==3.26.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.60.1
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-socket==0.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
svectordb==0.0.9
syrupy==4.8.1
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
timeago==1.0.14
together==1.3.13
toolz==1.0.0
tqdm==4.67.1
tree-sitter==0.24.0
tree-sitter-javascript==0.23.1
tree-sitter-php==0.23.11
tree-sitter-python==0.23.6
tree-sitter-typescript==0.23.2
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
voyageai==0.2.4
wcwidth==0.2.13
yarl==1.18.3
zstandard==0.23.0
